Adding screenshot script | Cleaning the code

+++ .config/hypr/scripts/charger_monitor.sh
@@ -5,11 +5,12 @@
 #----------------------------------------------------------
 
 #  INFO: CONFIGURATION ---
-SOUND_PLUG="/usr/share/sounds/freedesktop/stereo/power-plug.oga"
-SOUND_UNPLUG="/usr/share/sounds/freedesktop/stereo/power-unplug.oga"
+SOUND=true
+SOUND_PLUG="/usr/share/sounds/freedesktop/stereo/window-question.oga"
+SOUND_UNPLUG="/usr/share/sounds/freedesktop/stereo/window-attention.oga"
 # ------------------------
 
-#  INFO: Find the AC Adapter
+#  INFO: Find the AC Adapter ---
 AC_FILE=""
 for supply in /sys/class/power_supply/*; do
     if [ -f "$supply/type" ] && grep -q "Mains" "$supply/type"; then
@@ -18,7 +19,7 @@ for supply in /sys/class/power_supply/*; do
     fi
 done
 
-#  INFO: Error handler
+#  INFO: Error handler ---
 if [ -z "$AC_FILE" ]; then
     notify-send "Error" "Could not find AC Adapter."
     exit 1
@@ -26,25 +27,35 @@ fi
 
 last_state=$(cat "$AC_FILE")
 
+#  INFO: Main logic
 while true; do
     current_state=$(cat "$AC_FILE")
 
     if [ "$current_state" != "$last_state" ]; then
+        if "$SOUND"; then
+            # Create the lock file to silence the global notification script
+            touch /tmp/silence_notification_sound
 
-        #  INFO: Create the lock file to silence the global notification script
-        touch /tmp/silence_notification_sound
+            # Check the current charging state
+            if [ "$current_state" == "1" ]; then
+                paplay "$SOUND_PLUG"
+                notify-send -a "Power" "Power" "Charger Connected"
+            else
+                paplay "$SOUND_UNPLUG"
+                notify-send -a "Power" "Power" "Charger Disconnected"
+            fi
 
-        if [ "$current_state" == "1" ]; then
-            paplay "$SOUND_PLUG"
-            notify-send -a "Power" "Power" "Charger Connected"
+            # Wait a split second and remove the lock file
+            sleep 0.5
+            rm /tmp/silence_notification_sound
         else
-            paplay "$SOUND_UNPLUG"
-            notify-send -a "Power" "Power" "Charger Disconnected"
+            # Check the current charging state
+            if [ "$current_state" == "1" ]; then
+                notify-send -a "Power" "Power" "Charger Connected"
+            else
+                notify-send -a "Power" "Power" "Charger Disconnected"
+            fi
         fi
-
-        #  INFO:  Wait a split second and remove the lock file
-        sleep 0.5
-        rm /tmp/silence_notification_sound
 
         last_state=$current_state
     fi
